Add numeric conversions for counter-intuitive sort example
List::sort sorts *lexicographically*. This is a bit surprising when
the values are numbers -- however, we're keeping with Tcl's philosophy
where "everything is a string" and we can't really assume elements can
be coerced to numerical tyes.

For now, I think it's reasonable to have manual type constructors and
use List::map etc. in the construction of numeric lists.